Rename state entry points
