+++ src/app/contacts/states.js
@@ -1,7 +1,7 @@
-import create from './new/state';
-import edit from './edit/state';
-import list from './list/state';
-import show from './show/state';
+import create from './states/new/state';
+import edit from './states/edit/state';
+import list from './states/list/state';
+import show from './states/show/state';
 
 export default function($stateProvider) {
   'ngInject';
