+++ src/app/home/states.js
@@ -1,4 +1,4 @@
-import index from './index/state';
+import index from './states/index/state';
 
 export default function($stateProvider) {
   'ngInject';
